boba_community/hc-captcha-metafaucet/api/requirements.txt
python-dotenv
redis
web3>=5.29.2
captcha
